Added SignUp and Login views and urls

+++ orders/backend/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth import authenticate
 from django.core.validators import URLValidator
 from django.http import JsonResponse
 from django.shortcuts import render
@@ -7,50 +8,61 @@ from rest_framework.status import HTTP_401_UNAUTHORIZED
 from rest_framework.views import APIView
 from ujson import loads as load_json
 from yaml import load as load_yaml, Loader
+from rest_framework import generics, status
+from rest_framework.response import Response
+from rest_framework.request import Request
+from rest_framework.permissions import permission_classes
 
-from .serializers import ShopSerializer
+from .serializers import ShopSerializer, LoginSerializer, SignUpSerializer
 from .models import (Order, OrderItem, ProductInfo, ProductParameter, Parameter,
                      Product, Category, Shop, User)
 
 
 class PartnerUpdate(APIView):
-    def post(self, request, filename, *args, **kwargs):
-        filename = input('Enter file name: ')
+    def post(self, request, *args, **kwargs):
         if not request.user.is_authenticated:
             return JsonResponse({'detail': 'Authentication credentials were not provided.'}, status=HTTP_401_UNAUTHORIZED)
         if not request.user.type == 'shop':
             return JsonResponse({'detail': 'Only for shops'}, status=HTTP_401_UNAUTHORIZED)
 
-        with open(f'fixtures/{filename}', 'r') as stream:
-            data = load_yaml(stream, Loader=Loader)
-            shop, _ = Shop.objects.get_or_create(name=data['shop'], user_id=request.user.id)
-            for category in data['categories']:
-                category_obj, _ = Category.objects.get_or_create(id=category.id, name=category['name'])
-                category_obj.shops.add(shop.id)
-                category_obj.save()
+        url = request.data.get('url')
+        if url:
+            validate_url = URLValidator()
+            try:
+                validate_url(url)
+            except ValidationError:
+                return JsonResponse({'status': False, 'errors': 'Некорректная ссылка'}, status=400)
+            else:
+                stream = get(url).content
+                data = load_yaml(stream, Loader=Loader)
+                shop, _ = Shop.objects.get_or_create(name=data['shop'], user_id=request.user.id)
+                for category in data['categories']:
+                    category_obj, _ = Category.objects.get_or_create(id=category.id, name=category['name'])
+                    category_obj.shops.add(shop.id)
+                    category_obj.save()
 
-            ProductInfo.objects.filter(shop_id=shop.id).delete()
+                ProductInfo.objects.filter(shop_id=shop.id).delete()
 
-            for product in data['goods']:
-                product_obj, _ = Product.objects.get_or_create(name=product['name'], category_id=product['category_id'])
-                product_info_obj, _ = ProductInfo.objects.create(
-                        product_id=product_obj.id,
-                        shop_id=shop.id,
-                        model=product['model'],
-                        external_id=product['id'],
-                        quantity=product['quantity'],
-                        price=product['price'],
-                        price_rrc=product['price_rrc'],
-                    )
-                for name, value in product['parameters'].items():
-                    parameter_obj, _ = Parameter.objects.get_or_create(name=name)
-                    ProductParameter.objects.create(
-                            product_info_id=product_info_obj.id,
-                            parameter_id=parameter_obj.id,
-                            value=value
+                for product in data['goods']:
+                    product_obj, _ = Product.objects.get_or_create(name=product['name'], category_id=product['category_id'])
+                    product_info_obj, _ = ProductInfo.objects.create(
+                            product_id=product_obj.id,
+                            shop_id=shop.id,
+                            model=product['model'],
+                            external_id=product['id'],
+                            quantity=product['quantity'],
+                            price=product['price'],
+                            price_rrc=product['price_rrc'],
                         )
-                return JsonResponse({'status': True})
-            return JsonResponse({'status': False,  'errors': 'Не указаны все необходимые аргументы'})
+                    for name, value in product['parameters'].items():
+                        parameter_obj, _ = Parameter.objects.get_or_create(name=name)
+                        ProductParameter.objects.create(
+                                product_info_id=product_info_obj.id,
+                                parameter_id=parameter_obj.id,
+                                value=value
+                            )
+                    return JsonResponse({'status': True})
+                return JsonResponse({'status': False,  'errors': 'Не указаны все необходимые аргументы'})
 
 
 class ShopList(APIView):
@@ -66,4 +78,43 @@ class ShopDetails(APIView):
         return JsonResponse(serializer.data, safe=False)
 
 
+class SignUpView(generics.GenericAPIView):
+    serializer_class = SignUpSerializer
+
+    def post(self, request, *args, **kwargs):
+        serializer = self.serializer_class(data=request.data)
+        if serializer.is_valid:
+            user = serializer.validated_data
+            serializer.save()
+            response = {
+                "message": "User created successfuly",
+            }
+            return Response(data=response, status=status.HTTP_201_CREATED)
+
+        return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class LoginView(APIView):
+
+    def get(self, request):
+        content = {
+            'user': str(request.user),
+            'auth': str(request.auth)
+        }
+        return Response(data=content, status=status.HTTP_200_OK)
+
+    def post(self, request):
+        email = request.data.get('email')
+        password = request.data.get('password')
+        user = authenticate(email=email, password=password)
+
+        if user is not None:
+            response = {
+                'message': 'Login successful',
+                'email': user.email,
+                'Token': user.auth_token.key,
+            }
+            return Response(data=response, status=status.HTTP_200_OK)
+        else:
+            return Response(data={'message': 'Invalid credentials'}, status=status.HTTP_401_UNAUTHORIZED)
